WIP singleton

+++ src/main/y/concurrent/at-exit.hpp
@@ -10,9 +10,9 @@ namespace upsylon
     //! wrapper for libc atexit
     struct at_exit
     {
-        static const size_t stack_size = 64; //!< maximum number of register procedures
-        typedef void (*procedure)();         //!< a procedure to be called
-        typedef int    longevity;            //!< its longevity
+        static  const size_t stack_size = 64;   //!< maximum number of register procedures
+        typedef void       (*procedure)();      //!< a procedure to be called
+        typedef int16_t      longevity;         //!< its longevity
         //! register a new procedure
         /**
          The procedures are stored in a FIFO order (like atexit) for the
